[enigma2] Fix compiler warnings across multiple architectures (arm, mipsel, aarch64) - Fix implicit-fallthrough warnings with proper fall-through comments - Fix format-overflow warnings by using appropriately sized buffers - Fix sign-compare warnings with proper type casts - Fix unused-variable warnings by removing dead code (scan.cpp, elistbox.cpp) - Fix macro-redefined warnings with #ifndef guards (esize.h, python.h) - Guard DREAMBOX_DUAL_TUNER-only functions with #ifdef (dvbci.cpp) - Add -Wno-psabi to suppress ARM ABI change notes for STL iterator passing
cherry-picked > https://github.com/openatv/enigma2/commit/7757bff33aaccd9a3ca3d6c040fc9ae8f798805a

+++ lib/dvb/dvbtime.cpp
@@ -369,7 +369,7 @@ void eDVBLocalTimeHandler::updateTime(time_t tp_time, eDVBChannel *chan, int upd
 
 	if (m_time_debug)
 #ifdef GLIBC_64BIT_TIME_FLAGS
-		eDebug("[eDVBLocalTimerHandler] updateTime : %lld", tp_time);
+		eDebug("[eDVBLocalTimerHandler] updateTime : %lld", (long long)tp_time);
 #else
 		eDebug("[eDVBLocalTimerHandler] updateTime : %ld", tp_time);
 #endif
@@ -463,7 +463,7 @@ void eDVBLocalTimeHandler::updateTime(time_t tp_time, eDVBChannel *chan, int upd
 				time_difference = rtc_time - linuxTime;
 				if (m_time_debug)
 #ifdef GLIBC_64BIT_TIME_FLAGS
-					eDebug("[eDVBLocalTimerHandler] RTC to receiver time difference is %lld seconds.", linuxTime - rtc_time);
+					eDebug("[eDVBLocalTimerHandler] RTC to receiver time difference is %lld seconds.", (long long)(linuxTime - rtc_time));
 #else
 					eDebug("[eDVBLocalTimerHandler] RTC to receiver time difference is %ld seconds.", linuxTime - rtc_time);
 #endif
@@ -550,7 +550,7 @@ void eDVBLocalTimeHandler::updateTime(time_t tp_time, eDVBChannel *chan, int upd
 					new_diff = rtc - linuxTime; // Set enigma time to RTC.
 					if (m_time_debug)
 #ifdef GLIBC_64BIT_TIME_FLAGS
-						eDebug("[eDVBLocalTimerHandler] Update stored correction to %lld.  (Calculated against RTC time.)", rtc - tp_time);
+						eDebug("[eDVBLocalTimerHandler] Update stored correction to %lld.  (Calculated against RTC time.)", (long long)(rtc - tp_time));
 #else
 						eDebug("[eDVBLocalTimerHandler] Update stored correction to %ld.  (Calculated against RTC time.)", rtc - tp_time);
 #endif
